[Рефакторинг] Вынес элементы выбора текущего файла в отдельные компоненты для DIR и PCA Tools панелей
